backend/app/brain/f81f889b-15b9-41a6-bffd-57d89c75a45f/scratch/test_api.py
import requests

url = "http://localhost:8000/ai/chat/"
data = {"message": "serene getaway", "session_id": "test_session"}

try:
    response = requests.post(url, json=data)
    print("Status Code:", response.status_code)
    print("Response:", response.json())
except Exception as e:
    print("Error:", e)
