backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
groq
langchain<0.3.0
langchain-groq<0.2.0
langchain-community<0.3.0
sentence-transformers
numpy<2.0.0
faiss-cpu
pinecone
